Show the Drawer tutorial for upgrading users
With this change, users upgrading from Unity8 8.20 and below will
receive the "Long swipe from the left to open the Drawer" tutorial.
This should be the most important change that people used to Unity8 with
the Dash will need to know, so we can show it to them right away.